backend/pyproject.toml
[project]
name = "ai-robotics-book-backend"
version = "0.1.0"
description = "RAG-based chatbot backend for AI Humanoid Robotics Book"
requires-python = ">=3.10"
dependencies = [
    # FastAPI and ASGI server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.27.0",

    # Data validation
    "pydantic>=2.9.0",
    "pydantic-settings>=2.1.0",

    # Google Gemini SDK (embeddings)
    "google-generativeai==0.3.2",

    # OpenAI Agents SDK (for agent implementation)
    "openai-agents>=0.6.0",
    "openai>=1.0.0",

    # Vector database client
    "qdrant-client>=1.7.1",

    # PostgreSQL database
    "psycopg[binary]>=3.1.0",
    "sqlalchemy>=2.0.36",

    # Environment variables
    "python-dotenv==1.0.0",

    # Text processing
    "tiktoken>=0.5.2",
    "langchain>=0.1.5",
    "langchain-community>=0.0.16",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "httpx>=0.26.0",
    "black>=24.1.1",
    "flake8>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312', 'py313']

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
